Extend repl_send_goal to use the selected text as a goal even if not in a term
This patch extends `repl_send_goal` to use the selected text as a goal even
if the selection is outside a term, when the selection is not empty.

This is useful when the goal term is not defined as a constant but stored in an OCaml variable.
This frequently happens in s2n-bignum because it needs to prove a theorem that has to be dynamically built on the fly

```
let t = `1 + 1 = 2`;;

prove(t, ...);; (* Select 't' and press Alt+G ('repl_send_goal') *)
```