Let `repl_back_proof` highlight the previous tactic
This is a patch that updates `repl_back_proof` to highlight the previous tactic that has been executed before.
This is useful when one applied one or more `repl_back_proof`, saw the backtraced goal state, and wants to understand what was the last tactic that caused the backtraced goal state.
To achieve this, HolClient now has `tacticLocHistory` which is a list of VSCode locations.
This records text locations of tactics that are executed after goal setting so far.
This feature is enabled only when HOL Server is used, to add locations to `tacticLocHistory` only when `e` command was finished successfully.