Print the number of subgoals before '#' in HOL Client
This updates hol_client to print the number of subgoals as well as total subgoals before '#'.

The numbers are already printed as something like "val it : goalstack = 2 subgoals (2 total)"
when `g` or `e` is used in the basic HOL Light standard outputs.
However, it quickly goes behind a screen as soon as the printed goal state is too long or extra OCaml
commands like new definitions of tactics are executed.
This patch helps track the subgoal numbers.

To achieve this, I updated server2.ml so that the 'ready' message contains this information.
I reckoned that this was the easiest way to implement because otherwise an extra command for getting these
numbers must have been enqueued and its results must have been properly cached at somewhere.
Though, it makes server2.ml now dependent on the global variables that are already loaded on HOL Light.