Add line number directive, add 'Connect to Server' button
This patch includes addition of two different features:

1. Add line number directive to the beginning of every phrase sent to REPL

OCaml has a line number directive, e.g., `#100 "a.ml"` that is analogous to the
line directive of C/C++.
If this is attached to the beginning of a phrase that is sent to REPL, it
recognizes this and prints this when an exception is thrown & stack trace is enabled.

```
  # Printexc.record_backtrace true;;
..
  # #100 "a.ml"
    failwith "here";;

  Exception: Failure "here".
  Raised at Stdlib.failwith in file "stdlib.ml", line 29, characters 17-33
  Called from <unknown> in file "a.ml", line 100, characters 0-15
```

This is really useful in HOL Light if there is a tactic defined by a user and one wants
to understand why it is failing.

The official HOL Light does not support a line directive yet, however, because the old version
of camlp5 did not implement this, but since the latest 8.03.05 camlp5 now supports this directive
it is a matter of time that this will be fixed in HOL Light.
In fact, I already have a working version that supports this line directive:
  https://github.com/aqjune-aws/hol-light/tree/camlp5_8.03.05
After building this HOL Light with `make switch-5; eval $(opam env --set-switch); make`,
`hol.sh` will support this line directive.

2. Add 'Connect to Server' feature

From a few vscode users I received a question about how to connect to a running HOL Light server;
I found that it was indeed slighty confusing to find because I could connect through
'New HOL Light REPL Session'.
To address this, I added a 'shortcut' which is 'Connect to Server'.
Its implementation is analogous to `getTerminalWindow` because I started from its copy.